Implement addCopyPermalinkButtonToSidebar
On 2023-01-31, StoriesOnBoard updated the release view of a single
swimlane to include many useful features but removed a button to copy
a permament link of a card for the story map view. This extension adds
back this "Copy permalink" button to the sidebar when a user is viewing
a card inside a release view.